Add beta parameter to v_measure function
The beta parameter was unintentionally omitted. This PR adds the
missing parameter to the function definition and docstring. It also
fixes an error in v_measure_ct, where homogeneity and completeness
were mistakenly swapped. The error will not impact current users
as the computation is correct when beta=1.0. A unit test is also
added.